refactor: ログインページのLoginContentをコンポーネントとして分離し、SpeedInsightsを追加 (#806)
- frontend/app/login/page.tsx からLoginContentをfrontend/components/login-content.tsxへ切り出し
- @vercel/speed-insights を導入(Vercelホスティング時の計測用、現状のAWS ECS配信では未計測)
