+++ frontend/package.json
@@ -22,6 +22,7 @@
     "@radix-ui/react-radio-group": "^1.3.8",
     "@radix-ui/react-slot": "^1.1.0",
     "@vercel/analytics": "^1.3.1",
+    "@vercel/speed-insights": "^2.0.0",
     "class-variance-authority": "^0.7.0",
     "lucide-react": "^0.454.0",
     "next": "^16.0.4",
